+++ components/Posts/PostsOverview.vue
@@ -40,7 +40,7 @@
 
 <script>
 import IconsArrow from '@/components/Icons/IconsArrow'
-import getCategories from '@/assets/js/getCategories'
+import getCategories from 'assets/js/getUniqueCategories'
 
 export default {
   name: 'PostsOverview',
